requirements.txt
flask>=3.0.0
pymodbus>=3.6.0
python-dotenv>=1.0.1
